+++ iam-login-service/src/main/java/it/infn/mw/iam/config/oidc/OidcConfig.java
@@ -35,6 +35,7 @@ import org.springframework.security.web.authentication.AuthenticationSuccessHand
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
+import it.infn.mw.iam.api.openid_federation.FederatedOpRegistrationService;
 import it.infn.mw.iam.authn.AuthenticationSuccessHandlerHelper;
 import it.infn.mw.iam.authn.ExternalAuthenticationFailureHandler;
 import it.infn.mw.iam.authn.ExternalAuthenticationSuccessHandler;
@@ -49,6 +50,7 @@ import it.infn.mw.iam.authn.oidc.OIDCAuthenticationProvider;
 import it.infn.mw.iam.authn.oidc.OIDCAuthenticationToken;
 import it.infn.mw.iam.authn.oidc.OidcExceptionMessageHelper;
 import it.infn.mw.iam.authn.oidc.OidcTokenRequestor;
+import it.infn.mw.iam.authn.oidc.OpenIdFederationClientConfigurationService;
 import it.infn.mw.iam.authn.oidc.PlainAuthRequestUrlBuilder;
 import it.infn.mw.iam.authn.oidc.RestTemplateFactory;
 import it.infn.mw.iam.authn.oidc.service.OIDCProviderMetadataService;
@@ -61,6 +63,7 @@ import it.infn.mw.iam.core.IamThirdPartyIssuerService;
 import it.infn.mw.iam.core.client.IssuerService;
 import it.infn.mw.iam.core.jwk.IamJWKSetCacheService;
 import it.infn.mw.iam.persistence.repository.IamAccountRepository;
+import it.infn.mw.iam.persistence.repository.IamFederatedClientRepository;
 import it.infn.mw.iam.persistence.repository.IamTotpMfaRepository;
 
 @Configuration
@@ -82,8 +85,18 @@ public class OidcConfig {
   }
 
   @Bean
+  @Profile("!openid-federation")
   ClientConfigurationService clientConfigurationService(OidcProviderProperties oidcProperties) {
     return new DefaultClientConfigurationService(oidcProperties);
+  }
+
+  @Bean
+  @Profile("openid-federation")
+  ClientConfigurationService openIdFedclientConfigurationService(
+      OidcProviderProperties oidcProperties, Clock clock, IamFederatedClientRepository clientRepo,
+      FederatedOpRegistrationService federationRegistrationService) {
+    return new OpenIdFederationClientConfigurationService(oidcProperties, clock, clientRepo,
+        federationRegistrationService);
   }
 
   @Bean(name = "OIDCAuthenticationFilter")