+++ iam-login-service/src/test/java/it/infn/mw/iam/test/ext_authn/oidc/validator/OidcValidatorIntegrationTests.java
@@ -90,7 +90,7 @@ class OidcValidatorIntegrationTests extends OidcExternalAuthenticationTestsSuppo
   @BeforeEach
   void setup() {
     MockRestTemplateFactory tf = (MockRestTemplateFactory) restTemplateFactory;
-    tf.resetTemplate();
+    tf.resetServer();
   }
 
   @Test