+++ iam-login-service/src/test/java/it/infn/mw/iam/test/rcauth/RCAuthCertificateRequestorTests.java
@@ -82,7 +82,7 @@ class RCAuthCertificateRequestorTests extends RCAuthTestSupport {
   @BeforeEach
   void setup() {
     mockRtf = (MockRestTemplateFactory) rtf;
-    mockRtf.resetTemplate();
+    mockRtf.resetServer();
   }
 
   @Test
@@ -127,6 +127,6 @@ class RCAuthCertificateRequestorTests extends RCAuthTestSupport {
 
   void verifyMockServerCalls() {
     mockRtf.getMockServer().verify();
-    mockRtf.resetTemplate();
+    mockRtf.resetServer();
   }
 }