+++ iam-persistence/src/main/resources/db/migration/test/V100000___test_data.sql
@@ -33,7 +33,8 @@ INSERT INTO client_details (id, client_id, client_secret, client_name, dynamical
   (23, 'public-client-with-secret', 'secret', 'Public client with secret', false, 3600, 3600, 600, false, 'NONE', false, 600, CURRENT_TIMESTAMP(), 'optional', true),
   (24, 'pkce-s256-client', 'secret', 'Client with PKCE S256', false, 3600, 3600, 600, false, 'SECRET_BASIC', false, 600, CURRENT_TIMESTAMP(), 'S256', true),
   (25, 'pkce-plain-client', 'secret', 'Client with PKCE plain', false, 3600, 3600, 600, false, 'SECRET_BASIC', false, 600, CURRENT_TIMESTAMP(), 'plain', true),
-  (26, 'pkce-none-client', 'secret', 'Client with PKCE none', false, 3600, 3600, 600, false, 'SECRET_BASIC', false, 600, CURRENT_TIMESTAMP(), 'none', true);
+  (26, 'pkce-none-client', 'secret', 'Client with PKCE none', false, 3600, 3600, 600, false, 'SECRET_BASIC', false, 600, CURRENT_TIMESTAMP(), 'none', true),
+  (27, 'https://federated-client.com', 'secret', 'Federated client', false, 86400, 3600, 600, true, 'SECRET_BASIC', false, null, CURRENT_TIMESTAMP(), 'optional', true);
 
 UPDATE client_details SET client_description = 'implicit-flow-client description'
 WHERE id = 13 AND client_id = 'implicit-flow-client';
@@ -193,7 +194,8 @@ INSERT INTO client_scope (owner_id, scope) VALUES
   (25, 'openid'),
   (25, 'profile'),
   (26, 'openid'),
-  (26, 'profile');
+  (26, 'profile'),
+  (27, 'openid');
 
 INSERT INTO client_redirect_uri (owner_id, redirect_uri) VALUES
   (1, 'http://localhost:9090/iam-test-client/openid_connect_login'),
@@ -268,7 +270,8 @@ INSERT INTO client_grant_type (owner_id, grant_type) VALUES
   (23, 'client_credentials'),
   (24, 'authorization_code'),
   (25, 'authorization_code'),
-  (26, 'authorization_code');
+  (26, 'authorization_code'),
+  (27, 'client_credentials');
 
 INSERT INTO client_contact (owner_id, contact) VALUES
   (1, 'admin@example.com'),
@@ -1613,26 +1616,26 @@ INSERT INTO iam_account_group(account_id, group_id) VALUES
 (347,2);
 
 -- Owned clients
-insert into iam_account_client(id, account_id, client_id, creation_time) VALUES 
+INSERT INTO iam_account_client(id, account_id, client_id, creation_time) VALUES 
 (1, 200, 1, CURRENT_TIMESTAMP()),
 (2, 200, 2, CURRENT_TIMESTAMP()),
 (3, 199, 1, CURRENT_TIMESTAMP()),
 (4, 199, 2, CURRENT_TIMESTAMP());
 
 -- TOTP multi-factor secrets
-insert into iam_totp_mfa(active, secret, creation_time, last_update_time, account_id) VALUES
+INSERT INTO iam_totp_mfa(active, secret, creation_time, last_update_time, account_id) VALUES
 (true, '4S6Jqrvrv4eJDE9va5RkxaVjdjn5FCC958Gysekk1T7eel0pahEbp5iFLENeWZ4VKE1Dzit7AGowmOwD5i5Ma9VjnkdNMnU5WSK58w==', CURRENT_TIMESTAMP(), CURRENT_TIMESTAMP(), 1000);
 
 -- Client last used dates
-insert into client_last_used(client_details_id, last_used) VALUES
+INSERT INTO client_last_used(client_details_id, last_used) VALUES
 (1, '1994-03-21'),
 (2, '1994-03-20'),
 (3, '1994-03-19'),
 (4, '1994-03-23');
 
-insert into whitelisted_site(id, creator_user_id, client_id) values
+INSERT INTO whitelisted_site(id, creator_user_id, client_id) VALUES
 (1, 'admin', 'whitelist-client');
 
-insert into whitelisted_site_scope(owner_id, scope) values
+INSERT INTO whitelisted_site_scope(owner_id, scope) VALUES
 (1, 'openid'),
 (1, 'profile');